requirements.txt
streamlit==1.28.1
pandas==2.0.3
plotly==5.17.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.22
